+++ src/Relatus/Graphics/ImText.cs
@@ -5,14 +5,14 @@ namespace Relatus.Graphics
 {
     public static class ImText
     {
-        public static BetterSprite[] Create(float x, float y, string text, BMFont font, BMFontShader shader)
+        public static Sprite[] Create(float x, float y, string text, BMFont font, BMFontShader shader)
         {
             return Create(new Vector3(x, y, 0), text, font, shader);
         }
 
-        private static BetterSprite[] Create(Vector3 position, string text, BMFont font, BMFontShader shader)
+        private static Sprite[] Create(Vector3 position, string text, BMFont font, BMFontShader shader)
         {
-            BetterSprite[] sprites = new BetterSprite[text.Length];
+            Sprite[] sprites = new Sprite[text.Length];
 
             float xOffset = position.X;
 
@@ -21,7 +21,7 @@ namespace Relatus.Graphics
                 char character = text.Substring(i, 1).ToCharArray()[0];
                 BMFontCharacter characterData = font.GetCharacterData(character);
 
-                sprites[i] = BetterSprite.Create()
+                sprites[i] = Sprite.Create()
                     .SetTexture(font.GetPage(characterData.Page))
                     .SetSampleRegion(characterData.ImageRegion)
                     .SetRenderOptions(new RenderOptions()
