+++ src/Relatus/Graphics/SpriteElementsInstanced.cs
@@ -45,7 +45,7 @@ namespace Relatus.Graphics
             textureCoords = new VertexTexture[batchSize * 4];
         }
 
-        public override bool Add(BetterSprite sprite)
+        public override bool Add(Sprite sprite)
         {
             if (count >= BatchSize)
                 return false;
@@ -74,7 +74,7 @@ namespace Relatus.Graphics
             return true;
         }
 
-        public override DrawGroup<BetterSprite> ApplyChanges()
+        public override DrawGroup<Sprite> ApplyChanges()
         {
             if (!dataModified)
                 return this;
@@ -106,7 +106,7 @@ namespace Relatus.Graphics
             return this;
         }
 
-        public override DrawGroup<BetterSprite> Draw(Camera camera)
+        public override DrawGroup<Sprite> Draw(Camera camera)
         {
             if (dataModified)
                 throw new RelatusException("The sprite group was modified, but ApplyChanges() was never called.", new MethodExpectedException());
