+++ src/Relatus/Graphics/SpriteGroup.cs
@@ -1,6 +1,6 @@
 namespace Relatus.Graphics
 {
-    internal abstract class SpriteGroup : DrawGroup<BetterSprite>
+    internal abstract class SpriteGroup : DrawGroup<Sprite>
     {
         protected SpriteGroup(BatchExecution execution, uint batchSize) : base(execution, batchSize)
         {
